audio-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
httpx==0.27.2
faster-whisper==1.0.3
kokoro-onnx==0.4.6
onnxruntime-gpu==1.19.2
soundfile==0.12.1
chatterbox-tts
mcp==1.12.4
